MAINT: Restructure hierarchy of stainnorm
