+++ tiatoolbox/tools/__init__.py
@@ -1,2 +1,2 @@
 """Package defines tools for the toolbox"""
-from tiatoolbox.tools import stainnorm
+from tiatoolbox.tools import stainnorm, stain_utils
