+++ tiatoolbox/tools/stain_utils/__init__.py
@@ -1,0 +1,2 @@
+"""Package defines stain norm utils for the toolbox"""
+from tiatoolbox.tools import stain_extractor, abc_stain_extractor
