--- tiatoolbox/tools/stainnorm/__init__.py
@@ -1,2 +1,0 @@
-"""Package defines stain normalisation utilities for the toolbox"""
-from tiatoolbox.tools.stainnorm import reinhard_colour_normaliser, stain_normaliser
